+++ YouTube-client-app/src/app/core/components/header/header.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
-import { FilterShowService } from '../../../youtube/main-page/cards-list/services/filter-show-service.service';
+import { FilterShowService } from '../../../youtube/services/filter-show-service.service';
 import { LoginInfoComponent } from './login-info/login-info.component';
 import { SearchInputComponent } from './search-input/search-input.component';
 import { SettingsButtonComponent } from './settings-button/settings-button.component';
