+++ YouTube-client-app/src/app/youtube/directives/card-color.directive.ts
