+++ YouTube-client-app/src/app/youtube/main-page/cards-list/card-item/card-item.component.ts
@@ -2,8 +2,8 @@ import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 
 import { ButtonComponent } from '../../../../shared/button/button.component';
-import { CardColorDirective } from '../directives/card-color.directive';
-import { CardItemModel } from '../models/card-item.model';
+import { CardColorDirective } from '../../../directives/card-color.directive';
+import { CardItemModel } from '../../../models/card-item.model';
 
 @Component({
   selector: 'app-card-item',
