+++ YouTube-client-app/src/app/youtube/models/card-item.model.ts
